Move torchdynamo TestCase it its own file
[ghstack-poisoned]
