+++ test/dynamo/test_misc.py
@@ -16,6 +16,7 @@ from unittest.mock import patch
 import numpy as np
 import torch
 
+import torch._dynamo.test_case
 import torch._dynamo.testing
 import torch.onnx.operators
 from torch._dynamo import bytecode_transformation
@@ -34,7 +35,7 @@ def my_custom_function(x):
     return x + 1
 
 
-class MiscTests(torch._dynamo.testing.TestCase):
+class MiscTests(torch._dynamo.test_case.TestCase):
     def test_boolarg(self):
         def boolarg(aa, bb, flag):
             if flag:
@@ -2719,6 +2720,6 @@ class TestTracer(JitTestCase):
 
 
 if __name__ == "__main__":
-    from torch._dynamo.testing import run_tests
+    from torch._dynamo.test_case import run_tests
 
     run_tests()
