+++ test/dynamo/test_python_autograd.py
@@ -4,7 +4,8 @@ from typing import Callable, Dict, List, NamedTuple, Optional
 import torch
 
 import torch._dynamo
-from torch._dynamo.testing import CompileCounter, same, TestCase
+from torch._dynamo.test_case import run_tests, TestCase
+from torch._dynamo.testing import CompileCounter, same
 
 """
 This is an example of a pure-python version of autograd implemented by
@@ -283,6 +284,4 @@ class TestPythonAutograd(TestCase):
 
 
 if __name__ == "__main__":
-    from torch._dynamo.testing import run_tests
-
     run_tests()
