+++ test/dynamo/test_recompile_ux.py
@@ -6,10 +6,11 @@ import torch
 
 import torch._dynamo
 import torch._dynamo.config
+import torch._dynamo.test_case
 import torch._dynamo.testing
 
 
-class RecompileUxTests(torch._dynamo.testing.TestCase):
+class RecompileUxTests(torch._dynamo.test_case.TestCase):
     # TODO(whc) dynamo actualy recompiles one more time than the cache limit
     cache_limit = 1
