+++ test/dynamo/test_skip_non_tensor.py
@@ -4,10 +4,11 @@ from unittest.mock import patch
 import torch
 
 import torch._dynamo
+import torch._dynamo.test_case
 from torch._dynamo.testing import CompileCounter
 
 
-class SkipNonTensorTests(torch._dynamo.testing.TestCase):
+class SkipNonTensorTests(torch._dynamo.test_case.TestCase):
     def test_add_tensor1(self):
         def fn(a, b):
             return a + b
@@ -107,6 +108,6 @@ class SkipNonTensorTests(torch._dynamo.testing.TestCase):
 
 
 if __name__ == "__main__":
-    from torch._dynamo.testing import run_tests
+    from torch._dynamo.test_case import run_tests
 
     run_tests()
