+++ test/inductor/test_torchinductor.py
@@ -4023,6 +4023,6 @@ if HAS_CUDA:
 
 
 if __name__ == "__main__":
-    from torch._dynamo.testing import run_tests
+    from torch._dynamo.test_case import run_tests
 
     run_tests(needs="filelock")
